first crack at adding the connected_site script
In the public file, we check to see if the script url has been saved. If so, we apply it to the footer scripts.

Added function to grab the store,  check to see if the url and fragment have been applied, and if so, save the local variables for use.

Call the connect site endpoint to say it’s activated once it’s recognized for the first time.

Signed-off-by: Ryan Hungate <ryan@vextras.com>