new option: mailchimp-woocommerce-enable-landing-tracking
This adds a new option ( and UI ) to the plugin and disables the
`mailchimp_landing_site` cookie by default.

---

The current Mailchimp WordPress plugin will start adding cookies to all
page views the moment the plugin is activated.  You don't need to connect your
site to a Mailchimp account for this to happen.

Having a cookie on every page view will cause sites with CDNs to bypass the
CDN layer.  Activating this plugin immediately breaks the CDN cache.  That is
fine if a site owner decides that the functionality of having a cookie on every
page view is worth the trade off, but doing it by default is hurting the
performance of sites by default.

I work at Automattic, and we've noticed this problem on WooCommerce sites with
the Mailchimp plugin activated.

This has been brought up before - https://github.com/mailchimp/mc-woocommerce/issues/1225 - and there are code instructions for how to disable it - https://github.com/mailchimp/mc-woocommerce/wiki/Enable-or-Disable-Cookie-Filter - but most site owners are not going to get that far into it.  They expect a plugin
from a reliable source like Mailchimp to not hurt the performance of their
site by default.

Seven months ago Mailchimp indicated that they would look at this.  I haven't
seen any updates since then, so I'm offering this PR as a starting place to
preserve good site performance by default and make enabling the
`mailchimp_landing_site` cookie opt-in by the site owner.

Hopefully this plugin aligns with the style and arrangement that Mailchimp is
looking for in this plugin, but if not I'm happy to make what ever coding
adjustments they want.